pyproject.toml
[tool.poetry]
name = "teamidea-test-assignment"
version = "0.1.0"
description = ""
authors = ["Dmitry Samoylenko <dimsamoylenko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
